repair: remove callbacks
