+++ src/flamenco/repair/fd_repair_tool.c
@@ -54,7 +54,7 @@ repair_from_sockaddr( fd_repair_peer_addr_t * dst, uchar const * src ) {
   return 0;
 }
 
-static void
+static void FD_FN_UNUSED
 send_packet( uchar const * data,
              size_t        sz,
              fd_repair_peer_addr_t const * addr,
@@ -68,6 +68,104 @@ send_packet( uchar const * data,
     FD_LOG_WARNING(("sendto failed: %s", strerror(errno)));
   }
 }
+
+static void
+recv_shred(fd_shred_t const * shred, ulong shred_sz, fd_gossip_peer_addr_t const * from, fd_pubkey_t const * id, void * arg) {
+  (void)from;
+  (void)id;
+  (void)arg;
+  FD_LOG_NOTICE(( "shred variant=0x%02x sz=%lu slot=%lu idx=%u header_sz=0x%lx merkle_sz=0x%lx payload_sz=0x%lx",
+                  (uint)shred->variant, shred_sz, shred->slot, shred->idx, fd_shred_header_sz(shred->variant),
+                  fd_shred_merkle_sz(shred->variant), fd_shred_payload_sz(shred) ));
+}
+
+static void FD_FN_UNUSED
+deliver_fail_fun( fd_pubkey_t const * id,
+                         ulong               slot,
+                         uint                shred_index,
+                         void *              arg,
+                         int                 reason ) {
+  (void)arg;
+  FD_LOG_WARNING(( "repair_deliver_fail_fun - shred: %s, slot: %lu, idx: %u, reason: %d",
+                    FD_BASE58_ENC_32_ALLOCA( id ),
+                    slot,
+                    shred_index,
+                    reason ));
+}
+
+/* NOTE: since the removal of callbacks from repair_tile, this is now a similar but different
+   version of fd_repair_recv_clnt_packet in fd_repair_tile.c */
+/* Pass a raw client response packet into the protocol. addr is the address of the sender */
+int
+fd_repair_recv_clnt_packet( fd_repair_t *                 glob,
+                            uchar const *                 msg,
+                            ulong                         msglen,
+                            fd_repair_peer_addr_t const * src_addr,
+                            uint                          dst_ip4_addr FD_PARAM_UNUSED) {
+  glob->metrics.recv_clnt_pkt++;
+
+  FD_SCRATCH_SCOPE_BEGIN {
+    while( 1 ) {
+      fd_bincode_decode_ctx_t ctx = {
+        .data    = msg,
+        .dataend = msg + msglen
+      };
+
+      ulong total_sz = 0UL;
+      if( FD_UNLIKELY( fd_repair_response_decode_footprint( &ctx, &total_sz ) ) ) {
+        /* Solana falls back to assuming we got a shred in this case
+           https://github.com/solana-labs/solana/blob/master/core/src/repair/serve_repair.rs#L1198 */
+        break;
+      }
+
+      uchar * mem = fd_scratch_alloc( fd_repair_response_align(), total_sz );
+      if( FD_UNLIKELY( !mem ) ) {
+        FD_LOG_ERR(( "Unable to allocate memory for repair response" ));
+      }
+
+      fd_repair_response_t * gmsg = fd_repair_response_decode( mem, &ctx );
+      if( FD_UNLIKELY( ctx.data != ctx.dataend ) ) {
+        break;
+      }
+
+      switch( gmsg->discriminant ) {
+      case fd_repair_response_enum_ping:
+        /* TODO: make sure to add back handle_ping from repair tile in order to use */
+        //fd_repair_handle_ping( glob, &gmsg->inner.ping, src_addr, dst_ip4_addr );
+        break;
+      }
+
+      return 0;
+    }
+
+    /* Look at the nonse */
+    if( msglen < sizeof(fd_repair_nonce_t) ) {
+      return 0;
+    }
+    ulong shredlen = msglen - sizeof(fd_repair_nonce_t); /* Nonce is at the end */
+    fd_repair_nonce_t key = *(fd_repair_nonce_t const *)(msg + shredlen);
+    fd_needed_elem_t * val = fd_needed_table_query( glob->needed, &key, NULL );
+    if( NULL == val ) {
+      return 0;
+    }
+
+    fd_active_elem_t * active = fd_active_table_query( glob->actives, &val->id, NULL );
+    if( NULL != active ) {
+      /* Update statistics */
+      active->avg_reps++;
+      active->avg_lat += glob->now - val->when;
+    }
+
+    fd_shred_t const * shred = fd_shred_parse(msg, shredlen);
+    if( shred == NULL ) {
+      FD_LOG_WARNING(("invalid shread"));
+    } else {
+      recv_shred(shred, shredlen, src_addr, &val->id, glob->fun_arg);
+    }
+  } FD_SCRATCH_SCOPE_END;
+  return 0;
+}
+
 
 /* Convert a host:port string to a repair network address. If host is
  * missing, it assumes the local hostname. */
@@ -110,6 +208,7 @@ resolve_hostport(const char* str /* host:port */, fd_repair_peer_addr_t * res) {
 
   return res;
 }
+
 
 static int
 main_loop( int * argc, char *** argv, fd_repair_t * glob, fd_repair_config_t * config, volatile int * stopflag ) {
@@ -234,30 +333,6 @@ main_loop( int * argc, char *** argv, fd_repair_t * glob, fd_repair_config_t * c
   return 0;
 }
 
-static void
-recv_shred(fd_shred_t const * shred, ulong shred_sz, fd_gossip_peer_addr_t const * from, fd_pubkey_t const * id, void * arg) {
-  (void)from;
-  (void)id;
-  (void)arg;
-  FD_LOG_NOTICE(( "shred variant=0x%02x sz=%lu slot=%lu idx=%u header_sz=0x%lx merkle_sz=0x%lx payload_sz=0x%lx",
-                  (uint)shred->variant, shred_sz, shred->slot, shred->idx, fd_shred_header_sz(shred->variant),
-                  fd_shred_merkle_sz(shred->variant), fd_shred_payload_sz(shred) ));
-}
-
-static void
-deliver_fail_fun( fd_pubkey_t const * id,
-                         ulong               slot,
-                         uint                shred_index,
-                         void *              arg,
-                         int                 reason ) {
-  (void)arg;
-  FD_LOG_WARNING(( "repair_deliver_fail_fun - shred: %s, slot: %lu, idx: %u, reason: %d",
-                    FD_BASE58_ENC_32_ALLOCA( id ),
-                    slot,
-                    shred_index,
-                    reason ));
-}
-
 int main(int argc, char **argv) {
   fd_boot         ( &argc, &argv );
   fd_flamenco_boot( &argc, &argv );
@@ -281,10 +356,6 @@ int main(int argc, char **argv) {
 
   char const * my_addr = fd_env_strip_cmdline_cstr ( &argc, &argv, "--my_addr", NULL, ":1125");
   FD_TEST( resolve_hostport(my_addr, &config.intake_addr) );
-
-  config.deliver_fun = recv_shred;
-  config.clnt_send_fun = send_packet;
-  config.deliver_fail_fun = deliver_fail_fun;
 
   ulong seed = fd_hash(0, hostname, strnlen(hostname, sizeof(hostname)));
